requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
hdbscan>=0.8.27
umap-learn>=0.5.1
matplotlib>=3.4.0
seaborn>=0.11.0
geopy>=2.2.0
shapely>=1.8.0
geopandas>=0.10.0
